Split user services into three .NET 8 projects
