+++ api_write_user/Program.cs
@@ -1,8 +1,7 @@
 using Microsoft.AspNetCore.Mvc;
-using System.Linq;
 using MongoDB.Bson;
 using MongoDB.Driver;
-using UserApi.Models;
+using ApiWriteUser.Models;
 
 var builder = WebApplication.CreateBuilder(args);
 
@@ -17,40 +16,7 @@ builder.Services.AddSingleton(sp => sp.GetRequiredService<IMongoClient>().GetDat
 var app = builder.Build();
 
 var db = app.Services.GetRequiredService<IMongoDatabase>();
-var readCollection = db.GetCollection<User>("user_read");
 var writeCollection = db.GetCollection<User>("user_write");
-
-app.MapGet("/user/{id}", async (string id) =>
-{
-    if (!ObjectId.TryParse(id, out var objectId))
-        return Results.NotFound();
-
-    var user = await readCollection.Find(u => u.Id == objectId).FirstOrDefaultAsync();
-    return Results.Ok(new { response = user });
-});
-
-app.MapGet("/users", async (
-    [FromQuery] string? username,
-    [FromQuery] string? email,
-    [FromQuery] int pageSize = 10,
-    [FromQuery] int pageNumber = 1) =>
-{
-    var filter = Builders<User>.Filter.Empty;
-    if (!string.IsNullOrEmpty(username))
-        filter &= Builders<User>.Filter.Eq(u => u.Username, username);
-    if (!string.IsNullOrEmpty(email))
-        filter &= Builders<User>.Filter.Eq(u => u.Email, email);
-
-    if (pageSize <= 0) pageSize = 10;
-    if (pageNumber <= 0) pageNumber = 1;
-
-    var users = await readCollection.Find(filter)
-        .Skip((pageNumber - 1) * pageSize)
-        .Limit(pageSize)
-        .ToListAsync();
-
-    return Results.Ok(new { response = users, pageSize, pageNumber });
-});
 
 app.MapPost("/user", async (UserCreate request) =>
 {
@@ -98,27 +64,6 @@ app.MapDelete("/user/{id}", async (string id) =>
 
     var result = await writeCollection.DeleteOneAsync(u => u.Id == objectId);
     return Results.Ok(new { response = result.DeletedCount > 0 });
-});
-
-app.MapPost("/sync-manual", async () =>
-{
-    var unsynced = await writeCollection.Find(u => !u.IsSync).ToListAsync();
-    if (unsynced.Count > 0)
-    {
-        unsynced.ForEach(u =>
-        {
-            u.IsSync = true;
-            u.DateSynced = DateTime.UtcNow;
-        });
-
-        await readCollection.InsertManyAsync(unsynced);
-
-        var ids = unsynced.Select(u => u.Id).ToList();
-        var update = Builders<User>.Update.Set(u => u.IsSync, true);
-        await writeCollection.UpdateManyAsync(u => ids.Contains(u.Id), update);
-    }
-
-    return Results.Ok(new { response = true });
 });
 
 app.Run();
